[refactor] Create a file for the SpecialMethodsChecker in pylint.classes
